Calculator GUI

+++ I4SWD/Lecture0-0/Interfaces-2/Vehicles.Unit.Test/VehiclesUnitTest.cs
@@ -11,7 +11,7 @@ namespace Vehicles.Unit.Test
     [Author("NERK")]
     public class VehiclesUnitTest
     {
-        [TestCase(66)]
+        [TestCase((uint)66)]
         public void DieselEngine_Set_Get(uint a)
         {
             DieselEngine _suut = new DieselEngine(a);
@@ -21,7 +21,7 @@ namespace Vehicles.Unit.Test
             Assert.That(_suut.Throttle, Is.EqualTo(a / 2));
         }
 
-        [TestCase(88)]
+        [TestCase((uint)88)]
         public void GasEngine_Set_Get(uint a)
         {
             GasEngine _suut = new GasEngine(a);
@@ -31,7 +31,7 @@ namespace Vehicles.Unit.Test
             Assert.That(_suut.Throttle, Is.EqualTo(a / 2));
         }
 
-        [TestCase(122)]
+        [TestCase((uint)122)]
         public void ElectricEngine_Set_Get(uint a)
         {
             ElectricEngine _suut = new ElectricEngine(a);
